fix: force user to be on base to see mint button
