+++ src/components/MintDialog/elements/NativeMintButton.tsx
@@ -1,12 +1,10 @@
 import { FC } from 'react'
 import { Button } from '../../Button'
 import { ModalPage } from '../types'
-import { useChainId, useContract, useSwitchChain } from '@thirdweb-dev/react'
+import { useContract } from '@thirdweb-dev/react'
 import { TxDetails } from '../MintDialog'
 import { useMintDialogContext } from '../Context/useMintDialogContext'
-import { isProd } from '@/config/chain'
 
-import { base, baseGoerli } from 'viem/chains'
 interface NativeMintButtonProps {
   page: ModalPage
   setPage: React.Dispatch<ModalPage>
@@ -16,8 +14,6 @@ interface NativeMintButtonProps {
   setMintError: React.Dispatch<React.SetStateAction<any | null>>
 }
 
-const l2ChainId = isProd ? base.id : baseGoerli.id
-
 export const NativeMintButton: FC<NativeMintButtonProps> = ({
   page,
   setPage,
@@ -26,19 +22,13 @@ export const NativeMintButton: FC<NativeMintButtonProps> = ({
   setTxDetails,
   setMintError,
 }) => {
-  const switchChain = useSwitchChain()
-  const chainId = useChainId()
-
-  const wrongChain = chainId !== l2ChainId
   const isPending =
     page === ModalPage.NATIVE_MINTING_PENDING_TX ||
     page === ModalPage.NATIVE_MINT_PENDING_CONFIRMATION
   const { address } = useMintDialogContext()
   const { data: contract, isLoading } = useContract(address)
 
-  return wrongChain ? (
-    <Button onClick={() => switchChain(l2ChainId)}>Switch to Base</Button>
-  ) : (
+  return (
     <Button
       className="!flex text-black text-lg font-medium w-full justify-between rounded-lg"
       disabled={isPending || isLoading}
